+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.27.0] - 2025-06-01
+MR #29
+
+### Added
+- Alerts API route
+- Alerts API visualisations
+- Cache for all API routes
+- Mapping for transport lines icons
+
+### Changed
+- Traffic API route to use the new cache
+
 ## [0.23.0] - 2025-05-30
 MR #25